Adding category annotations and editing category controller

+++ UltimateSecuritySurvey/Controllers/categoryController.cs
@@ -9,20 +9,36 @@ using UltimateSecuritySurvey.Models;
 
 namespace UltimateSecuritySurvey.Controllers
 {
+    /// <summary>
+    /// This controller to Displays Categories and to manage them
+    /// </summary>
     public class categoryController : Controller
     {
+
+       
         private SecuritySurveyEntities db = new SecuritySurveyEntities();
 
         //
         // GET: /category/
 
+        /// <summary>
+        /// This method to list the existing categories.
+        /// Here is also a view bag to display delete errors
+        /// </summary>
+
         public ActionResult Index()
         {
+            ViewBag.DeleteError = (TempData["DeleteError"]) ?? string.Empty;
             return View(db.QuestionCategories.ToList());
         }
 
         //
         // GET: /category/Details/5
+
+        /// <summary>
+        /// This method finds one category and shows it individually with details.
+        /// </summary>
+        /// <param name="id">Primary of Question Category</param>
 
         public ActionResult Details(string id = null)
         {
@@ -37,6 +53,10 @@ namespace UltimateSecuritySurvey.Controllers
         //
         // GET: /category/Create
 
+        /// <summary>
+        /// This method to go to the create view
+        /// </summary>
+
         public ActionResult Create()
         {
             return View();
@@ -45,6 +65,10 @@ namespace UltimateSecuritySurvey.Controllers
         //
         // POST: /category/Create
 
+        /// <summary>
+        /// This method to add new categories
+        /// </summary>
+        /// <param name="questioncategory">Question category object from textbox</param>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create(QuestionCategory questioncategory)
@@ -62,6 +86,10 @@ namespace UltimateSecuritySurvey.Controllers
         //
         // GET: /category/Edit/5
 
+        /// <summary>
+        /// This method to edit view, finds the category details
+        /// </summary>
+
         public ActionResult Edit(string id = null)
         {
             QuestionCategory questioncategory = db.QuestionCategories.Find(id);
@@ -74,6 +102,10 @@ namespace UltimateSecuritySurvey.Controllers
 
         //
         // POST: /category/Edit/5
+        /// <summary>
+        /// This method to edit the category details, cant change the id
+        /// </summary>
+        /// <param name="questioncategory">Existing category object</param>
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -90,6 +122,9 @@ namespace UltimateSecuritySurvey.Controllers
 
         //
         // GET: /category/Delete/5
+        /// <summary>
+        /// This method to show chosen category details and confirm delete
+        /// </summary>
 
         public ActionResult Delete(string id = null)
         {
@@ -104,14 +139,44 @@ namespace UltimateSecuritySurvey.Controllers
         //
         // POST: /category/Delete/5
 
+        /// <summary>
+        /// This method to delete the existing category
+        /// </summary>
+        /// <param name="id">The primary key of the category</param>
+
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
             QuestionCategory questioncategory = db.QuestionCategories.Find(id);
-            db.QuestionCategories.Remove(questioncategory);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (CheckIfQuestionsExistForCategory(questioncategory.categoryId))
+            {
+                db.QuestionCategories.Remove(questioncategory);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["DeleteError"] = "Cannot delete this Category because related questions exist!!";
+                return RedirectToAction("Index");
+            }
+         }
+        /// <summary>
+        /// This method to Check if this category has questions in question table, in a case of delete violation
+        /// </summary>
+        /// <param name="id">The chosen category id</param>
+
+        private bool CheckIfQuestionsExistForCategory(string id)
+        {
+            Question question = db.Questions.Where(x => x.categoryId == id).First();
+            if (question == null)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
         protected override void Dispose(bool disposing)
